fix: 修复合并后的 5 处编译错误
- toKline：收盘时间缺失时回落开盘时间，不再误把 Long 当返回值
- WebSocketListener.onClosing 的 reason 在 OkHttp 4 为非空 String
- 补 awaitClose 导入（kotlinx.coroutines.channels）
- rawLimit 用 Long 计算再收窄，避免 Int.coerceAtMost(Long) 类型不符
- DefaultWsHosts 在 data.remote.ws 包，容器导入路径写错

+++ app/src/main/java/com/waxilo/marketmonitor/data/repository/MarketRepositoryImpl.kt
@@ -182,7 +182,9 @@ class MarketRepositoryImpl(
     ): KlinePage {
         val base = interval.baseInterval ?: error("周期缺少基础周期")
         // 聚合需要整数倍根数，多取一点避免最后一桶总是不够
-        val rawLimit = (limit * interval.aggregateRatio).coerceAtMost(BinanceMarketApi.MAX_KLINE_LIMIT.toLong()).toInt()
+        val rawLimit = (limit.toLong() * interval.aggregateRatio)
+            .coerceAtMost(BinanceMarketApi.MAX_KLINE_LIMIT.toLong())
+            .toInt()
         val raw = api.klines(id.market, id.symbol, base.apiCode, rawLimit, startTime, endTime)
         val now = System.currentTimeMillis()
         val stamped = raw.map { if (it.closeTime > now) it.copy(closed = false) else it }
